+++ src/main/java/com/masstransitproject/crosstown/newid/NewIdGenerator.java
@@ -61,28 +61,28 @@ public class NewIdGenerator {
 			sequence = _sequence++;
 		}
 
-		return new NewId(_a, _b,  _c,(_d | sequence));
+		return new NewId(_a, _b, _c, (_d | sequence));
 	}
 
 	void UpdateTimestamp(long tick) {
-		
-//		ByteBuffer buffer = ByteBuffer.allocate(8);
-//	    buffer.putLong(tick);
-//	    
-//	    buffer.flip();
-//
-//
-//	    _a = buffer.getInt();
-//	    _b = buffer.getInt();
-	    
+
+		// ByteBuffer buffer = ByteBuffer.allocate(8);
+		// buffer.putLong(tick);
+		//
+		// buffer.flip();
+		//
+		//
+		// _a = buffer.getInt();
+		// _b = buffer.getInt();
 
 		_lastTick = tick;
 		_sequence = 0;
-		
-		 //Java requires this be an unsigned shift in order to be consistent with the c#
-		 _a = (long) (tick >>> 32);
-		 
-		 //Java requires this be stored as a long so the shifting isn't lossy
-		 _b = (long) (tick << 32 >>> 32) ; 
+
+		// Java requires this be an unsigned shift in order to be consistent
+		// with the c#
+		_a = tick >>> 32;
+
+		// Java requires this be stored as a long so the shifting isn't lossy
+		_b = tick << 32 >>> 32;
 	}
 }
